requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.1
flake8==6.0.0
psycopg2-binary==2.9.6
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.3.1
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
